requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
curtsies==0.4.1
cwcwidth==0.1.8
datasets==2.10.1
Deprecated==1.2.13
dill==0.3.6
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.3.0
huggingface-hub==0.13.1
idna==3.4
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
pyarrow==11.0.0
pycparser==2.21
PyGithub==1.58.0
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
wrapt==1.15.0
xxhash==3.2.0
yarl==1.8.2
